+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-02T09:26:56.098Z",
+    "2026-02-02T10:53:00.225Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {